ucx: implement GASNET_UCX_MAX_MEDIUM env var
This commit implements (and documents) the environment variable
`GASNET_UCX_MAX_MEDIUM` for controlling the LUB Medium value for ucx
conduit (and PSHM, when enabled).  The use of "MAX", rather then
"LUB", in the name is to align with the established usage in other
conduits, such as aries.  A new `--with-ucx-max-medium` configure
argument is added as a means to set the _default_ value, which is
now 8KiB minus overheads.

Additionally, while updating the max payload queries for the new
variable buffer size, they have also been modified to eliminate the
use of a function call to obtain the AM Medium header size.  Instead a
manifest constant is used, and a new `gasneti_static_assert()` ensures
that we know if its value becomes out-of-date.