ucx: split Request and Reply send pools
This commit updates the logic in `gasnetc_am_req_get()` to allocate
send buffers for Requests and Replies from distinct pools, with
associated changes to how the pools are initialized and destroyed.

Initially this simply increases the available resources.  Longer-term
it is anticipated that this split may be necessary to prevent certain
deadlock scenarios.