ofi: exit improvements for bug 4360
This commit makes the following improvements to the logic implementing
`gasnet_exit()` in ofi-conduit:

1. All fixed-duration `alarm()` calls in the exit code have been
subsumed into a smaller number of alarm intervals with durations
related to the `GASNET_EXITTIMEOUT` family of environment variables.
This ensures the user has the means to increase the intervals, and
thus addresses the original problem reported as bug 4360.

2. Increased `GASNETC_DEFAULT_EXITTIMEOUT_MAX` to 8 minutes, based on
recent `MPI_Finalize()` timings on Summit, extrapolated (roughly) to
full system scale.