Update gasnet_extended_fwd.h
This commit updates `extended-ref/gasnet_extended_fwd.h` to contain
a slightly more complete description of the settings related to
use and configuration of the AM-based reference extended.

At the same time, the use of `#if 0` has been replaced with use of
`//` comments to ease search for non-default values of any given
setting.  This is mostly in the updated section, but not exclusively.

The conduits with their own `gasnet_extended_fwd.h` have been
updated as appropriate.