Implement gasneti_{quit,fatal}_signal_rcvd
This commit adds global variables `gasneti_quit_signal_rcvd` and
`gasneti_fatal_signal_rcvd`, which are set by the default signal
handler upon receipt of a corresponding signal.

This will improve the ability for conduit code to avoid "risky"
activity in the event of exits from the default `SIGQUIT` handler, as
well as eliminating the need for conduits to implement a fatal signal
callback for the sole purpose of detecting fatal signal context.

Note that this does not provide a mechanism for a `gasnet_exit()`
implementation to detect that it has been invoked from a *client's*
handler for `SIGQUIT`.