Add other/gasnet_refexit.c
This commit adds an `other/gasnet_refexit.c` source file and includes
it in the `gasnet_core.c` of the three conduits which utilize a common
approach to exit handling (ibv, ofi and ucx).

This file is "empty" of declarations and definitions, but will grow
(and the `gasnet_core.c` files shrink) as portions of the exit
handling are factored.