ofi: proper "stable" provider selection
This commit sets the provider name field in the `hints` passed to
`fi_getinfo()` calls after the initial one.  This ensures that when
multiple providers satisfy the other hints, we really do use the
one we identified based on the first `fi_getinfo()` call.  Prior
attempts to do this based on setting the domain in `hints` have
been seen to be insufficient.