ofi: clone the reference extended
This commit creates `gasnet_extended.c` and `gasnet_extended_fwd.h`
in `ofi-conduit/`, based on the ones in `extended-ref/`.

All RMA in ofi-conduit is still using the "amref" implementation, and
currently does so at the expense of an additional function call
relative to prior to this commit.  Future work will replace this with
native libfabric RMA.