WIP - ofi: native RMA
[REWORD prior to de-WIP to ensure all changes are covered]

This implementation is a simple forward-port of the extended API
implementation of the ofi-conduit in GASNet-1.  In particular there
are only "bulk" and "non-bulk" (`DEFER` and `NOW`) behaviors available
for `lc_opt`, with the others mapped to `NOW`

Until now the conduit-specifc RDMA put and get logic from GASNet-1 has
remained unreachable in `gasnet_ofi.c` and underwent only minor
updates for changes to types.  So, this commit makes some additional
updates to resolve latent bugs.

1. Update completion handling for NBI RMA operations to correctly
interact with NBI access regions (aop) and their offset-by-one
counters.

2. Update the handling of memory registration keys for non-bulk
(EVENT_NOW) Puts to accommodate the split of the aux and client
segments.

3. Repair logic for non-bulk (EVENT_NOW) Puts to use `gasneti_wait()`
and an internal EOP.  This resolves some efficiency issues, including
over synchronization in the NBI case while simultaneously addressing a
correctness issue with aops.

4. Remove some dead code.