ofi: currently only one user segment is supported
This commit makes a second (or subsequent) call to create a segment
an error (returning `GASNET_ERR_RESOURCE`).

The memory key handling in ofi-conduit is currently only a binary
aux-vs-client distinction.  Therefore, more than a single call to
either `gex_Segment_Attach()` or `gex_Segment_Create()` results in
a segment which cannot be correctly accessed if bound to an EP.

Since ofi-conduit does not yet support multiple endpoints, and
use of `gex_Segment_Attach()` is still the only supported means to
bind a segment to the primordial endpoint, there is currently also
no valid use case (outside of coverage/conformance tests) to create
multiple segments with ofi-conduit.