Improved thread cleanup toward bug 4287
The is work toward Bug 4287 "Deferred cleanup of thread-specific data"
which is motivated by the near-future addition of some exit-time
handling of nbi_ff.  The changes include:

1. Identify leaked threaddata instances with a new `is_undead` field.

2. Use the undead status to ensure that if `gasnete_free_threaddata()`
leaks the nbi_ff aop, then a re-run of `gasnete_free_threaddata()` for
deferred cleanup will behave correctly (not erroneously repeating the
adjustment of the offset counters).

3. Moves mutex manipulation outward to ensure thread cleanup (including
marking as a zombie) is completed atomically.