ofi: deploy gasneti_finalize_all_nbi_ff
With the recent addition of a native extended API implementation to
ofi-conduit, the "RDMADISSEM" barrier has become the default.  The
implementation of that barrier uses the "nbi_ff" facility to "sink"
all of its RMA Puts.  This can result in RMA operations still "in
flight" from the perspective of the ofi provider at exit time
(especially the Puts in a barrier immediately preceding the exit).
In the case of the `verbs;ofi_rxm` provider, this prevents a clean
shutdown (a failure to `fi_close(...gasnetc_ofi_domainfd...)` in
particular).

This commit uses the new `gasneti_finalize_all_nbi_ff()` call to get
an array of the aops for all of the "nbi_ff" operations which have not
been synced prior to exit.  It then performs a bounded-time wait for
their completion before proceeding with the exit.