ofi: send AM Medium nbytes in-band
This commit claims a byte of padding in the AM header to transmit the
length of AM Medium payloads, replacing the use of 8-bytes of "remote
CQ data" with use of otherwise wasted space.  The user's `nbytes` is
encoded using just a single byte as difference between it and the full
length of the message sent by `fi_send()`.  Since the recipient is
provided this length, it can reconstruct `nbytes`.

A similar change is possible for Longs using a packed format.  For
Longs using RDMA for payload transfer, it is possible to add a 32-bit
field to the header to use in place of the 64-bit CQ data.  However,
any such changes are being deferred until target-side reassembly can
be implemented which may require more significant changes.