ofi: cleanup fabric address fetch
This commit replaces the `GET_{RDMA,AM_{REQUEST,REPLY}}_DEST()`
family of macros with a more accurately named `gasnetc_fabric_addr()`.

In addition to the name change is the elimination of explicit
small-integer constants in favor of an enum, together with static
assertions to ensure the insertion order matches the constants used in
queries.  This will ease maintenance, such as the possible future
addition of more ofi-level endpoints (for instance, as may be
required/desired for Long payloads with target-side reassembly).