ofi: complete gasneti_EP_PublishBoundSegment()
This commit completes ofi-conduit's implementation of
`gasneti_EP_PublishBoundSegment()`, allowing it to pass `testsegment`
without a special case (which is thus removed in this commit).  This
brings the implementation of segments up to the same level as other
conduits which lack multiple endpoints.  To be specific, this work
removes the assumption under `FI_MR_SCALABLE` that the primordial
endpoint is associated with the primordial segment (which always
has `mr_key == 1` ).

The primary change is to stop special casing `FI_MR_SCALABLE` with
respect to key management, and in particular removing the
short-circuiting of the conduit-specific publish hook and therefore
maintaining real tables of the registration keys.

The prior "simplification" for `FI_MR_SCALABLE` was based on the
assumption that aux and client segments (with fixed keys 0 and 1,
respectively) were the only memory registrations.  That remains true
(for the moment) only for `EVERYTHING` segment mode, but is not true
otherwise due to `gex_Segment_Create()`.  With the introduction of
Memory Kinds, that won't be true of any segment mode.

Additional changes include dropping the unused registration of the aux
segment in `EVERYTHING` segment mode and thus avoiding adding a new
table for that case.  For the moment, `EVERYTHING` segment mode is
still a special case, but addition of memory kinds (if it includes
this mode) will likely reduce or eliminate the differences.

Finally, the `FI_WRITE` and `FI_READ` macros have been revised to
eliminate needless duplication between the `BASIC` and `SCALABLE`
memory registration paths.

This work adds a table look up to the critical path for RDMA injection
under `FI_MR_SCALABLE && !GASNET_SEGMENT_EVERYTHING`.  However,
multi-endpoint support will soon require at least that amount of
overhead.