ofi: revise scalable memory registration
This commit revises and completes ofi-conduit's implementation of
memory registration.  The "revision" allows effective use of the
ability to specify the memory registration key under `FI_MR_SCALABLE`,
while the "completion" enables ofi-conduit to pass `testsegment`
without a special case (which is thus removed in this commit).  This
brings the implementation of segments up to the same level as other
conduits which lack support for multiple endpoints, while making
progress toward such support.

The primary change is to ensure that under `FI_MR_SCALABLE` a remote
memory registration key is always computable (and thus never needs to
be stored).  This is accomplished by deferring memory registration
until `gex_EP_BindSegment()`, when we can use the endpoint index to
select a "computable" key (the computation is simply adding 1, to
reserve 0 for the aux segment's key).  While this alone is not
sufficient for support of multiple endpoints, it does ensure that no
storage of remote registration keys will be required under
`FI_MR_SCALABLE`.

Additional changes include (1) dropping the unused registration of the aux
segment in `EVERYTHING` segment mode, and (2) work toward logical
separation of the multiple provider properties which comprise
`FI_MR_SCALABLE`.

Finally, the `FI_WRITE` and `FI_READ` macros have been revised to
eliminate needless duplication between the `BASIC` and `SCALABLE`
memory registration paths.  Additional cleanups have been deferred
to the work on multi-endpoint support.