ofi: implement multi-endpoint support
This commit implements multiple endpoint support in ofi-conduit.  As
with ibv and ucx, the current default `GASNET_MAXEPS` is 33 and only
`GEX_EP_CAPABILITY_RMA` is supported.

The main change is to update the remote key and address lookups to use
the remote EP index, instead of `in_auxseg`.  An otherwise invalid EP
index value of -1 is used to identify the remote aux segment, and keys
and arrays are setup so this is continuous with the valid indices
(avoiding branches on `in_auxseg` or the index which replaces it).

In the case of `FI_MR_PROV_KEY` (part of `FI_MR_BASIC`) the remote
registration keys are communicated in the conduit-dependent portions
of `gex_Segment_Attach()` and `gex_EP_PublishBoundSegment()`.  In the
absence of `FI_MR_PROV_KEY`, keys are computed and never stored.

In the case of `FI_VIRT_ADDR` (part of `FI_MR_BASIC`) the ofi-level
RMA function use virtual addresses.  In the other (`FI_MR_SCALABLE`)
case, offsets from the base of each registration are used.  The
conduit-independent portions of `gex_Segment_Attach()` and
`gex_EP_PublishBoundSegment()` already communicate and store the
segment bounds.  So no additional storage is needed.

Additionally, this commit adds the appropriate conduit-specific case
to `testtmpair` and updates documentation.