REORDER-AT-REBASE: ofi: prevent misuse of GASNETC_OFI_HAS_MR_SCALABLE
AT REBASE - make this commit come before the "real work".

This commit ensures that `GASNETC_OFI_HAS_MR_SCALABLE` cannot be
(mis)used in preprocessor directives.  By adding an explicit cast to
the definition, the value as a C expression remains unchanged, but use
in a preprocessor expression becomes a syntax error.