Increase kinds test coverage via testlarge/small
This commit expands the options related to the (intentionally
undocumented, for now) `-cuda-uva` and `-hip` options to `testlarge`
and `testsmall`.  This permits the testing of all four combinations
of device versus host memory at the local and remote end of the RMA.

A final "done." message has also been added to these two tests.

Additionally, `harness.conf` has been updated to exercise all four
combinations and to collect the benchmark numbers for each.