pmi-spawner: default to Cray PMI on an HPE Cray EX
This commit updates the PMI configure probe to default to
`--with-pmi-version=cray` on the HPE Cray EX (Shasta) platform,
and improves the failure message to distinguish a user-provided
version from a defaulted one.

Note that this does NOT default to enabling pmi-spawner on an HPE Cray
EX -- it just changes the default version when it is enabled.