ofi: force PSM2_MULTI_EP=1
This commit ensures that `PSM2_MULTI_EP=1` in the environment prior to
potentially initializing libpsm2 via a call to `fi_getinfo()`.

This addresses the scenario in which some other agent (such as Intel
MPI) has set `PSM2_MULTI_EP=0` but is *not* using the psm2 provider,
including the possibility that the "other agent" is an environment
module, the user's `.profile`, or similar.

If/when trying to allow for use of psm2 provider by ofi-conduit and
another runtime, then it is possible that libpsm2 has been initialized
prior to ofi-conduit initialization, in which case this is
insufficient (too late).