WIP - rmabuf experiment
This commit is based on reading libfabric and MPICH sources and a public
slide presentation from Intel.  It reproduces the Linux "rmabuf" logic
which is apparently used to register Intel GPU memory instead of the
use of a `/sys/kernel/mm/memory_peers` driver as with GDR and
ROCmRDMA.

If this proves useful, then it needs a proper abstraction.