ofi: generalize "aux" context bits
This commit generalizes the concept of using the low bits of an
`op_context` pointer.  Specifically, it is no longer assumed that the
`alc` property of RMA operation contexts will be the only use of these
bits.  The `alc` bit is given a name (rather than use of the constant
`1`) and is explicitly tested for in the `aux` bits rather than
assuming `alc` and `aux` are interchangeable.

No use of the second available bit is added in this commit, but a
future commit to do so will no longer break `alc` handling.