NPAM: initial GEX_FLAG_IMMEDIATE_COMMIT
This commit provides the specification of `GEX_FLAG_IMMEDIATE_COMMIT`
and a initial implementation.

The implementation is "opt-in", such that a conduit needs to define
new per-operation identifiers in their `gasnet_core_fwd.h` in either
of two cases:

  1. When using the reference NPAM over an FPAM (specifically
     `gasnetc_{Request,Reply}{Medium,Long}V()`) which honors
     `GEX_FLAG_IMMEDIATE` for network AMs (in-nbrhd uses a
     different path in the reference).
  2. When providing a native Commit which honors the new flag.

In the absence of the opt-in, preprocessor logic provides a Commit
definition of roughly `(gasnetc_commitFoo(args), 0)`.  This means no
change of return type needed in current conduits.  Additionally,
future use of this flag in clients is amenable to compiler
optimization when there is no possibility of "immediate failure" of
Commit.

No tests yet *use* this flag.  However, `testgasnet.c` has been
updated to verify existence and uniqueness (w/i a relevant group) of
the new flag.