Factor threaddata initialization logic
This commit factors an idiom used in every implementation (reference
and native) of `gasnete_init()` in which the threaddata is initialized
for the thread calling `gex_Client_Init()`.  The factored version of
the logic is written to be correct if called multiple times, with a
near-term expectation of use of thread-specific data prior to the
client's call to `gex_Client_Init()`.