ibv: progress thread init hook for bug 4706
This commit implements and documents hooks (provided via the new
`gex_Client_InitHooks()`) which enable a client to run code within
ibv-conduit's send and receive progress threads when they start.  This
is expected to provide the client with the ability to control CPU
affinity and memory affinity of these threads.