Provide client-facing GASNET_{RCV,SND}_THREAD
This commit provides preprocessor identifiers `GASNET_RCV_THREAD` and
`GASNET_SND_THREAD` to allow a client to predicate actions which
depend of the presence of the corresponding support in a client.  In
particular, this allows a client runtime to avoid setting the
same-named environment variables when support was disabled at
configure time, thus avoiding a warning about the conflict.
Additionally, it allows more forward-looking client code than would
use of `GASNET_CONDUIT_IBV` to control conditional compilation of
logic associated with progress threads.