fixup - ofi: add distance info to GASNET_OFI_LI..
The `README` for ofi-conduit has been updated to reflect that the
`GASNET_OFI_LIST_DEVICES` support for distance ranking can be
provider-dependent.  The currently known dependencies are enumerated.

It turns out that there are no providers for which GASNet supports
libfabric prior to 1.7 *and* PCI bus information is needed to perform
the distance ranking:

+ `cxi` needs PCI bus info, added in libfabric 1.7, but this provider
  is not supported until 1.15
+ `psm2` through at least libfabric 1.22.0 provides neither PCI bus
  info nor any device/domain name other than "psm2" and thus cannot
  support hwloc-based distance measurements in any current release.
+ All other providers I could access for testing can use the
  device/domain names as far back as we support the provider.