README: update PSHM env var docs
This commit updates the documentation for the environment variable
`GASNET_SUPERNODE_MAXSIZE` to introduce the term "nbhrd" and to
mention `gex_System_QueryNbrhdInfo()` ahead of `gasnet_getNodeInfo()`.

The term "nbrhd" is added to the docs for `GASNET_PSHM_BARRIER_HIER`.

An out-of-date reference to "node" as a synonym for "process" has
also been removed.