Implement and document GASNET_USE_PSHM_SINGLETON
As described in new documentation, this commit introduces a new
default behavior for a "singleton" nbrnd (one containing a single
process) which avoids use of shared memory allocation (since no actual
cross-mapping will be performed), instead falling back on
`gasneti_mmap()`.  This new environment variable can be set to a true
value in order to to restore the prior behavior in which shared memory
is allocated in singleton nbrhds, despite never being shared.