Support libhugetlbfs in non-default locations
This commit adds support in `configure` for `libhugetlbfs` installed
outside the compiler's default header and library search paths.  This
includes two new configure options: `--with-hugetlbfs-home` and
`--enable-hugetlbfs-rpath`, along with a `HUGETLBFS_HOME` environment
variable as a lower-precedence alternative the former.

Additionally, the probe for `libhugetlbfs` support has been revised to
use `GASNET_TRY_CACHE_LINK` in place of nested `AC_CHECK_HEADERS` and
`AC_CHECK_FUNCS`.  In addition to caching, this change removes the
_unused_ `HAVE_HUGETLBFS_H` and `HAVE_HUGETLBFS_UNLINKED_FD` from
`gasnet_config.h`.

Finally, I've taken advantage of this opportunity to expand the
configure probe to include _all_ of the libhugetlbfs calls used in
GASNet-EX.